[frontend] ContainerStixDomainObjects migration to tsx (#4328)
