+++ opencti-platform/opencti-front/src/utils/hooks/useLocalStorage.ts
@@ -58,6 +58,8 @@ export interface UseLocalStorageHelpers {
     symbol?: string;
     original?: number;
   }) => void;
+  handleToggleTypes: (type: string) => void;
+  handleClearTypes: () => void;
   handleAddProperty: (field: string, value: unknown) => void;
   handleChangeView: (value: string) => void;
   handleToggleTypes: (type: string) => void;
